Trigger enable preview hint on 'not.supported.in.source' error
The javac shipped with NB may support features which are no longer
preview, while still being preview on the target JDK of the project.

To trigger the enable-preview hint in that scenario, it would have
to listen to 'compiler.err.feature.not.supported.in.source' as
workaround.

This isn't technically correct but will hopefully do the right thing
most of the time.